sorted files and made ikon use more sexy
